requirements.txt
gspread
google-auth
streamlit
pandas
plotly
kaleido